requirements.txt
comtypes
JPype1